+++ Source/Core/TimeInterval.js
@@ -118,6 +118,7 @@ define([
      * @param {TimeInterval} [left] The first interval.
      * @param {TimeInterval} [right] The second interval.
      * @param {Function} [dataComparer] A function which compares the data of the two intervals.  If ommitted, reference equality is used.
+     *
      * @returns {Boolean} <code>true</code> if left and right are equal, <code>false</code> otherwise.
      */
     TimeInterval.equals = function(left, right, dataComparer) {
@@ -127,8 +128,8 @@ define([
                (left.isEmpty && right.isEmpty ||
                 left.isStartIncluded === right.isStartIncluded &&
                 left.isStopIncluded === right.isStopIncluded &&
-                left.start.equals(right.start) &&
-                left.stop.equals(right.stop) &&
+                JulianDate.equals(left.start, right.start) &&
+                JulianDate.equals(left.stop, right.stop) &&
                 (left.data === right.data ||
                  (defined(dataComparer) && dataComparer(left.data, right.data))));
     };
@@ -142,6 +143,7 @@ define([
      * @param {TimeInterval} [left] The first TimeInterval.
      * @param {TimeInterval} [right] The second TimeInterval.
      * @param {Number} epsilon The epsilon to use for equality testing.
+     * @param {Function} [dataComparer] A function which compares the data of the two intervals.  If ommitted, reference equality is used.
      *
      * @returns {Boolean} <code>true</code> if left and right are within the provided epsilon, <code>false</code> otherwise.
      *
@@ -158,8 +160,8 @@ define([
                (left.isEmpty && right.isEmpty ||
                 left.isStartIncluded === right.isStartIncluded &&
                 left.isStopIncluded === right.isStopIncluded &&
-                left.start.equalsEpsilon(right.start, epsilon) &&
-                left.stop.equalsEpsilon(right.stop, epsilon) &&
+                JulianDate.equalsEpsilon(left.start, right.start, epsilon) &&
+                JulianDate.equalsEpsilon(left.stop, right.stop, epsilon) &&
                 (left.data === right.data ||
                  (defined(dataComparer) && dataComparer(left.data, right.data))));
     };
@@ -216,14 +218,14 @@ define([
 
         if (otherStart.greaterThanOrEquals(thisStart) && thisStop.greaterThanOrEquals(otherStart)) {
 
-            isStartIncluded = (!otherStart.equals(thisStart) && otherIsStartIncluded) || (thisIsStartIncluded && otherIsStartIncluded);
+            isStartIncluded = (!JulianDate.equals(otherStart, thisStart) && otherIsStartIncluded) || (thisIsStartIncluded && otherIsStartIncluded);
 
             isStopIncluded = thisIsStopIncluded && otherIsStopIncluded;
 
             outputData = defined(mergeCallback) ? mergeCallback(this.data, other.data) : this.data;
 
             if (thisStop.greaterThanOrEquals(otherStop)) {
-                isStopIncluded = isStopIncluded || (!otherStop.equals(thisStop) && otherIsStopIncluded);
+                isStopIncluded = isStopIncluded || (!JulianDate.equals(otherStop, thisStop) && otherIsStopIncluded);
                 return new TimeInterval(otherStart, otherStop, isStartIncluded, isStopIncluded, outputData);
             }
 
@@ -233,13 +235,13 @@ define([
 
         if (otherStart.lessThanOrEquals(thisStart) && thisStart.lessThanOrEquals(otherStop)) {
 
-            isStartIncluded = (otherStart.equals(thisStart) === false && thisIsStartIncluded) || (thisIsStartIncluded && otherIsStartIncluded);
+            isStartIncluded = (JulianDate.equals(otherStart, thisStart) === false && thisIsStartIncluded) || (thisIsStartIncluded && otherIsStartIncluded);
 
             isStopIncluded = thisIsStopIncluded && otherIsStopIncluded;
 
             outputData = defined(mergeCallback) ? mergeCallback(this.data, other.data) : this.data;
             if (thisStop.greaterThanOrEquals(otherStop)) {
-                isStopIncluded = isStopIncluded || (otherStop.equals(thisStop) === false && otherIsStopIncluded);
+                isStopIncluded = isStopIncluded || (JulianDate.equals(otherStop, thisStop) === false && otherIsStopIncluded);
                 return new TimeInterval(thisStart, otherStop, isStartIncluded, isStopIncluded, outputData);
             }
 
@@ -286,6 +288,8 @@ define([
      * @memberof TimeInterval
      *
      * @param {TimeInterval} [right] The right hand side Cartesian.
+     * @param {Function} [dataComparer] A function which compares the data of the two intervals.  If ommitted, reference equality is used.
+     *
      * @returns {Boolean} <code>true</code> if they are equal, <code>false</code> otherwise.
      */
     TimeInterval.prototype.equals = function(other, dataComparer) {
@@ -300,6 +304,8 @@ define([
      *
      * @param {TimeInterval} [right] The right hand side Cartesian.
      * @param {Number} epsilon The epsilon to use for equality testing.
+     * @param {Function} [dataComparer] A function which compares the data of the two intervals.  If ommitted, reference equality is used.
+     *
      * @returns {Boolean} <code>true</code> if they are within the provided epsilon, <code>false</code> otherwise.
      *
      * @exception {DeveloperError} epsilon is required and must be a number.